manager: split code of manager into several submodules
Split code of manager into several submodules to improve readability.
